chore: refactor folders and finish fixing errors
