+++ src/university/src/Bitfields/Bitfield.ts
