+++ src/university/src/Bitfields/User.ts
