+++ src/university/src/RestManager.ts
@@ -1,10 +1,11 @@
-import { RestPayload, RestRequest } from "../../../../rest/rest.ts";
-import { DiscordHTTPResponseCodes } from "../../../../types/codes/http_response_codes.ts";
-import { FileContent } from "../../../../types/discordeno/file_content.ts";
-import { Collection } from "../../../../util/collection.ts";
-import { BASE_URL, IMAGE_BASE_URL, USER_AGENT } from "../../../../util/constants.ts";
-import { camelize, delay } from "../../../../util/utils.ts";
-import Client from "../Client.ts";
+import { RestPayload, RestRequest } from "../../rest/rest.ts";
+import { ModifyChannel } from "../../types/channels/modify_channel.ts";
+import { DiscordHTTPResponseCodes } from "../../types/codes/http_response_codes.ts";
+import { FileContent } from "../../types/discordeno/file_content.ts";
+import { Collection } from "../../util/collection.ts";
+import { BASE_URL, IMAGE_BASE_URL, USER_AGENT } from "../../util/constants.ts";
+import { camelize, delay } from "../../util/utils.ts";
+import Client from "./Client.ts";
 
 export class RestManager {
   /** The bot client this is managing for. */
@@ -40,6 +41,10 @@ export class RestManager {
   version: number;
   /** The authorization secret used if you are using a proxy rest server. */
   authorization?: string;
+  /** The queue to editing the channel names or topics since they have special subratelimits. */
+  editChannelNameTopicQueue = new Map<bigint, EditChannelRequest>();
+  /** Whether or not the edit channel queue is currently being processed. */
+  editChannelProcessing = false;
 
   constructor(client: Client) {
     this.client = client;
@@ -512,3 +517,17 @@ export class RestManager {
 }
 
 export default RestManager;
+
+export interface EditChannelRequest {
+  amount: number;
+  timestamp: number;
+  channelId: bigint;
+  items: {
+    channelId: bigint;
+    reason?: string;
+    options: ModifyChannel;
+    resolve: (channel: unknown) => void;
+    // deno-lint-ignore no-explicit-any
+    reject: (error: any) => void;
+  }[];
+}
